Create 2015-12-10-Deliverable-D72-Available
We've submitted the second WP7 (D7.2) deliverable and it is already available on our Deliverables Page at: http://www.cyclone-project.eu/deliverables.html.
D7.2 - Overlay with focus on Component Manager is the accompanying release note for the first CYCLONE demonstrator set up, which was set as an internal milestone of the project and carried out during the General Assembly meeting in Berlin on 7-Oct-2015 (M10 of the project). We are also planning to make the demonstrator available very soon.
The focus of the demonstration was twofold.  At first, we aimed at demonstrating the integration of specific security components with SlipStream Component Manager; these components are the Federated Identity Provider and the Distributed Logging System. Then, we demonstrated the integration of the Federated Identity Provider to extend two different applications: a common WordPress deployment and one of the CYCLONE Bioinformatics use cases (Securing human bio	medical data) described in D3.1 [1]. SlipStream owns a predominant role in this first demonstrator acting as Component Manager.